requirements.txt
humanfriendly
passlib
PyYAML
kubernetes==3.0.0
